refactor: make family-only into server component
